refactor: consolidate error type definitions into unified module
Consolidate all error types (Database, Provider, Embedding) into a single unified error module at `/packages/core/src/shared/errors/database.ts`. This eliminates duplication and provides consistent error handling patterns across the codebase.

Changes:
- Consolidated 10 error types from 3 separate files into single module
- Converted EmbeddingDataParseError from Error class to Data.TaggedError for consistency
- Updated 18 files to import from unified error module
- Removed duplicate error definitions from providers/types.ts
- Fixed all EmbeddingDataParseError constructor calls to use object syntax

Benefits:
- Single source of truth for all error type definitions
- Consistent error field patterns (message, cause, errorCode, etc.)
- Easier to maintain and extend error types
- Better error handling with Effect's tagged error system

Fixes #141
